Tiny fixes in test_feature project
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@137 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/tests/test_feature/source/testexecutescript.cpp
@@ -44,8 +44,8 @@ bool TestExecuteScript()
 const char *GetCurrentDir(char *buf, size_t size)
 {
 #ifdef _MSC_VER
-	return _getcwd(buf, size);
-#elsif __APPLE__
+	return _getcwd(buf, (int)size);
+#elif defined(__APPLE__)
 	return getcwd(buf, size);
 #else
 	return "";
